requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
flask-mail
psycopg2-binary
python-dotenv
gunicorn
itsdangerous
openai>=1.0.0
google-generativeai>=0.3.0
requests
